Rename inner `update_supply_chain_information` to `config`.
