+++ update_supply_chain_information/config/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for update_supply_chain_information project.
+ASGI config for config project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,8 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault(
-    "DJANGO_SETTINGS_MODULE", "update_supply_chain_information.settings"
-)
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
 
 application = get_asgi_application()
