+++ update_supply_chain_information/config/local_settings.py
@@ -1,6 +1,6 @@
-from update_supply_chain_information.settings import *
+from config.settings import *
 
-ROOT_URLCONF = "update_supply_chain_information.local_urls"
+ROOT_URLCONF = "config.local_urls"
 MIDDLEWARE += ["debug_toolbar.middleware.DebugToolbarMiddleware"]
 INSTALLED_APPS += ["debug_toolbar"]
 INTERNAL_IPS = ["127.0.0.1"]
