+++ update_supply_chain_information/config/local_urls.py
@@ -1,6 +1,6 @@
 import debug_toolbar
 
-from update_supply_chain_information.urls import *
+from config.urls import *
 
 urlpatterns += [
     path("__debug__/", include(debug_toolbar.urls)),
