+++ update_supply_chain_information/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = update_supply_chain_information.settings
+DJANGO_SETTINGS_MODULE = config.settings
 python_files = tests.py test_*.py *_tests.py
 
 ; RemovedInDjango41Warning: 'webpack_loader' defines default_app_config = 'webpack_loader.apps.WebpackLoaderConfig'. Django now detects this configuration automatically. You can remove default_app_config.
