+++ update_supply_chain_information/supply_chains/test/test_middleware.py
@@ -2,13 +2,13 @@ import pytest
 from django.test import Client
 from django.urls import reverse
 
-from update_supply_chain_information import settings
+from config import settings
 
 
 def test_correct_middleware_exists():
     assert settings.MIDDLEWARE == [
         "django.middleware.security.SecurityMiddleware",
-        "update_supply_chain_information.middleware.add_cache_control_header_middleware",
+        "config.middleware.add_cache_control_header_middleware",
         "django.contrib.sessions.middleware.SessionMiddleware",
         "django.middleware.common.CommonMiddleware",
         "django.middleware.csrf.CsrfViewMiddleware",
